pyproject.toml
[project]
name = "final-pratice-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.8",
    "python-decouple>=3.8",
    "rich>=14.1.0",
]
